refactor(cli): split remote command module [OPE-367]
